Updated tests and decimal precision

+++ src/main/java/com/harsain/RPNCalculator/Expression/Expression.java
@@ -2,11 +2,12 @@ package com.harsain.RPNCalculator.Expression;
 
 import com.harsain.RPNCalculator.OperatorEnum;
 
+import java.math.BigDecimal;
 import java.util.Collection;
 import java.util.List;
 
 public interface Expression {
-    List<Double> interpret();
+    List<BigDecimal> interpret();
 
     Collection<? extends Expression> getOperands();
